problemset-175887/discounts/requirements.txt
Django==4.0.4
djangorestframework==3.13.1
django-countries==7.3.2
